Log 'missing values' diamond import errors in a more Sentry-friendly manner
This gathers up all messages for a given field (clarity, cut_aliases, etc)
into the same Sentry group.  The list of values that were missing appears
under 'Additional Data'.  Events can also be filtered on backend, which is
now a tag.

As a side effect, to see the list of missing values in a log file, a new
formatter will be needed.  E.g, format = ... %(missing_values)